backend/pyproject.toml
[project]
name = "moviereservationsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.24",
]
